src/referee_gloves_detector/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
DateTime==5.1
filelock==3.12.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
pandas==2.0.2
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
zope.interface==6.0
